+++ src/toisto/model/quiz/quiz_factory.py
@@ -139,13 +139,18 @@ class QuizFactory:
         )
 
 
-def grammatical_quiz_types(concept1: Concept, concept2: Concept) -> tuple[QuizType, ...]:
-    """Return the quiz types to change the grammatical category of concept1 into that of concept2."""
+def grammatical_quiz_types(concept1: Concept, concept2: Concept, max_quizzes: int = 2) -> tuple[QuizType, ...]:
+    """Return the quiz types to change the grammatical category of concept1 into that of concept2.
+
+    For example, to change "I am" into "they are" would mean changing the grammatical number from singular to plural
+    and changing the grammatical person from first person to third person. To prevent the quiz from becoming too
+    complex ("Give the affirmative past tense plural third person...") we limit the number of quiz types.
+    """
     quiz_types = []
     for category1, category2 in zip_longest(concept1.grammatical_categories(), concept2.grammatical_categories()):
         if category1 != category2 and category2 in GRAMMATICAL_QUIZ_TYPES:
             quiz_types.append(GRAMMATICAL_QUIZ_TYPES[category2])
-    return tuple(quiz_types)
+    return tuple(quiz_types) if len(quiz_types) <= max_quizzes else tuple()
 
 
 def paired_leaf_concepts(concept: Concept) -> Iterable[tuple[Concept, Concept]]: